Added objective-based operation:
The robot will no longer act on a rotating fixed state machine, but instead read objectives.

Each objective has a corresponding point, a name, and an action to take.
Objectives are sequenced in order of the "sequence" number.

SCAN => Scan with the side camera
TRACK => Track with the radar
NONE => Do nothing
DONE => Quit

TODO:
* should add a "precise" point flag to an objective
* implement TRACK
* Implement "action information" such that the corresponding action is given
  all necessary information to perform it's duty (e.g. num items, duration, whatever)